Dev frontend

+++ employee-management-backend/src/admin/admin.controller.ts
@@ -8,19 +8,27 @@ import {
   Put,
   UseGuards,
   ParseIntPipe,
+  Query,
+  Req,
+  NotFoundException,
 } from '@nestjs/common';
 import { AdminService } from './admin.service';
 import { CreateAdminDto } from './dto/create-admin.dto';
 import { UpdateAdminDto } from './dto/update-admin.dto';
 import { JwtAuthGuard } from 'src/jwt/jwt-auth.guard';
+import type { Request } from 'express';
+import { HandleLeaveException } from 'src/interceptors/handleLeave.interceptor';
 
-@Controller('admin')
+@Controller('admins')
 @UseGuards(JwtAuthGuard)
 export class AdminController {
   constructor(private readonly adminService: AdminService) {}
   @Get()
-  async findAll() {
-    const data = await this.adminService.findAll();
+  async findAll(
+    @Query('page', ParseIntPipe) page = 1,
+    @Query('rowsPerPage', ParseIntPipe) rowsPerPage = 5,
+  ) {
+    const data = await this.adminService.findAll(page, rowsPerPage);
     return {
       status: 200,
       message: 'success get all admins',
@@ -65,7 +73,12 @@ export class AdminController {
   }
 
   @Delete(':id')
-  async remove(@Param('id', ParseIntPipe) id: number) {
+  async remove(@Param('id', ParseIntPipe) id: number, @Req() req: Request) {
+    const loggedInUser = req?.user as { id: number };
+
+    if (loggedInUser?.id === id) {
+      throw new HandleLeaveException(`You cannot delete yourself!`, 400);
+    }
     const data = await this.adminService.remove(id);
 
     return {
